Added 'layoutPanel' control in Booking form (this control is used to draw the different rooms that the users will be able to click and select)
Implemented a hit registration system for 'layoutPanel' to detect whether a room has been selected by the user.

Implemented a collision check for the user's checkin and checkout date to determine whether a room is available during the range of dates that the customer wants to book.

Implemented various validations to the form.